Add files via upload
This project was developed for the Moldovan winery Vinia Traian. JSC VINIA TRAIAN company was founded in 1975 and is one of the largest wine producers in the region with the protected geographical indication “Valul lui Traian” (Trajan's Wave) - the southernmost wine-growing region of Moldova, historically located between the two valleys of the Traian Wave.